backend/app/requirements.txt
python-dotenv
fastapi
sqlalchemy
sqlalchemy-utils
alembic
async-exit-stack 
async-generator
requests 
sendgrid
boto3
pytz
pyjwt
passlib[bcrypt]
python-multipart
pillow
gunicorn
uvicorn
uvloop
httptools
psycopg2-binary
pytest
pytest-cov
pytest-asyncio

pycodestyle
autopep8
